+++ product-api/src/main/java/br/com/rocha/productapi/ProductApiApplication.java
@@ -1,0 +1,13 @@
+package br.com.rocha.productapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ProductApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ProductApiApplication.class, args);
+	}
+
+}
