+++ product-api/src/test/java/br/com/rocha/productapi/ProductApiApplicationTests.java
@@ -1,0 +1,13 @@
+package br.com.rocha.productapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ProductApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
