--- src/main/java/br/com/rocha/JavaBackend/JavaBackendApplication.java
@@ -1,13 +1,0 @@
-package br.com.rocha.JavaBackend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class JavaBackendApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(JavaBackendApplication.class, args);
-	}
-
-}
