--- src/main/java/br/com/rocha/JavaBackend/dto/UserDTO.java
@@ -1,74 +1,0 @@
-package br.com.rocha.JavaBackend.dto;
-
-import java.util.Date;
-
-import br.com.rocha.JavaBackend.model.User;
-
-public class UserDTO {
-
-    private String name;
-    private String cpf;
-    private String adress;
-    private String email;
-    private String telephone;
-    private Date dateRegistry;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getCpf() {
-        return cpf;
-    }
-
-    public void setCpf(String cpf) {
-        this.cpf = cpf;
-    }
-
-    public String getAdress() {
-        return adress;
-    }
-
-    public void setAdress(String adress) {
-        this.adress = adress;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getTelephone() {
-        return telephone;
-    }
-
-    public void setTelephone(String telefone) {
-        this.telephone = telefone;
-    }
-
-    public Date getDateRegistry() {
-        return dateRegistry;
-    }
-
-    public void setDateRegistry(Date dateRegistry) {
-        this.dateRegistry = dateRegistry;
-    }
-
-    public static UserDTO convert(User user) {
-        UserDTO userDTO = new UserDTO();
-        userDTO.setName(user.getName());
-        userDTO.setCpf(user.getCpf());
-        userDTO.setAdress(user.getAdress());
-        userDTO.setEmail(user.getEmail());
-        userDTO.setTelephone(user.getTelephone());
-        userDTO.setDateRegistry(user.getDateRegistry());
-        return userDTO;
-    }
-}
