--- src/main/java/br/com/rocha/JavaBackend/repository/UserRepository.java
@@ -1,16 +1,0 @@
-package br.com.rocha.JavaBackend.repository;
-
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import br.com.rocha.JavaBackend.model.User;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    User findByCpf(String cpf);
-
-    List<User> queryByNameLike(String name);
-}
