--- src/test/java/br/com/rocha/JavaBackend/JavaBackendApplicationTests.java
@@ -1,13 +1,0 @@
-package br.com.rocha.JavaBackend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class JavaBackendApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
