+++ user-api/src/main/java/br/com/rocha/userapi/UserApiApplication.java
@@ -1,0 +1,13 @@
+package br.com.rocha.userapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class UserApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(UserApiApplication.class, args);
+	}
+
+}
