+++ user-api/src/main/java/br/com/rocha/userapi/controller/UserController.java
@@ -1,0 +1,128 @@
+package br.com.rocha.userapi.controller;
+
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+import br.com.rocha.userapi.dto.UserDTO;
+import br.com.rocha.userapi.service.UserService;
+
+
+
+@RestController
+public class UserController{
+
+    @Autowired
+    private UserService userService;
+
+    @GetMapping("/user/")
+    public List<UserDTO> getUsers() {
+
+        List<UserDTO> users = userService.getAll();
+        return users;
+    }
+
+    @GetMapping("/user/{id}")
+    UserDTO findById(@PathVariable Long id){
+        return userService.findById(id);
+    }
+
+    @PostMapping("/user")
+    UserDTO newUser(@RequestBody UserDTO userDTO){
+        return userService.save(userDTO);
+    }
+
+    @GetMapping("user/cpf/{cpf}")
+    UserDTO findByCpf(@PathVariable String cpf){
+        return userService.findByCpf(cpf);
+    }
+
+    @DeleteMapping("/user/{id}")
+    UserDTO delete(@PathVariable Long id){
+        return userService.delete(id);
+    }
+
+    @GetMapping("/user/search")
+    public List<UserDTO> queryByName(@RequestParam(name = "name", required = true)String name){
+        return userService.queryByName(name);
+    }
+}
+/* @RequestMapping("/user")
+public class UserController {
+
+    public static List<UserDTO> users = new ArrayList<UserDTO>();
+
+    @PostConstruct
+    public void initiateList() {
+        UserDTO userDTO = new UserDTO();
+        userDTO.setName("Eduardo");
+        userDTO.setCpf("123");
+        userDTO.setAdress("Rua a");
+        userDTO.setEmail("eduardo@email.com");
+        userDTO.setTelephone("1234-3454");
+        userDTO.setDateRegistry(new Date());
+
+        UserDTO userDTO2 = new UserDTO();
+        userDTO2.setName("Luiz");
+        userDTO2.setCpf("456");
+        userDTO2.setAdress("Rua b");
+        userDTO2.setEmail("luiz@email.com");
+        userDTO2.setTelephone("1234-3454");
+        userDTO2.setDateRegistry(new Date());
+
+        UserDTO userDTO3 = new UserDTO();
+        userDTO3.setName("Bruna");
+        userDTO3.setCpf("678");
+        userDTO3.setAdress("Rua c");
+        userDTO3.setEmail("bruna@email.com");
+        userDTO3.setTelephone("1234-3454");
+        userDTO3.setDateRegistry(new Date());
+
+        users.add(userDTO);
+        users.add(userDTO2);
+        users.add(userDTO3);
+    }
+
+    @GetMapping("/users")
+    public List<UserDTO> getUsers() {
+        return users;
+    }
+
+    @GetMapping("users/{cpf}")
+    public UserDTO getUsersFilter(@PathVariable String cpf) {
+
+        for (UserDTO userFilter : users) {
+            if (userFilter.getCpf().equals(cpf)) {
+                return userFilter;
+            }
+        }
+        return null;
+    }
+
+    @PostMapping("/newUser")
+    public UserDTO postUser(@RequestBody UserDTO userDTO) {
+        userDTO.setDateRegistry(new Date());
+        users.add(userDTO);
+        return userDTO;
+    }
+
+    @DeleteMapping("/{cpf}")
+    public boolean remove(@PathVariable String cpf) {
+
+        for (UserDTO userFilter : users) {
+            if (userFilter.getCpf().equals(cpf)) {
+                users.remove(userFilter);
+                return true;
+            }
+        }
+        return false;
+    }
+}
+ */
