+++ user-api/src/main/java/br/com/rocha/userapi/model/User.java
@@ -1,0 +1,88 @@
+package br.com.rocha.userapi.model;
+
+import java.util.Date;
+
+import br.com.rocha.userapi.dto.UserDTO;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+
+@Entity
+public class User {
+
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+    private String name;
+    private String cpf;
+    private String adress;
+    private String email;
+    private String telephone;
+    private Date dateRegistry;
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getCpf() {
+        return cpf;
+    }
+
+    public void setCpf(String cpf) {
+        this.cpf = cpf;
+    }
+
+    public String getAdress() {
+        return adress;
+    }
+
+    public void setAdress(String adress) {
+        this.adress = adress;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getTelephone() {
+        return telephone;
+    }
+
+    public void setTelephone(String telephone) {
+        this.telephone = telephone;
+    }
+
+    public Date getDateRegistry() {
+        return dateRegistry;
+    }
+
+    public void setDateRegistry(Date dateRegistry) {
+        this.dateRegistry = dateRegistry;
+    }
+
+    public static User convert(UserDTO userDTO) {
+        User user = new User();
+        user.setName(userDTO.getName());
+        user.setCpf(userDTO.getCpf());
+        user.setAdress(userDTO.getAdress());
+        user.setEmail(userDTO.getEmail());
+        user.setTelephone(userDTO.getTelephone());
+        user.setDateRegistry(userDTO.getDateRegistry());
+        return user;
+    }
+}
