+++ user-api/src/main/java/br/com/rocha/userapi/repository/UserRepository.java
@@ -1,0 +1,18 @@
+package br.com.rocha.userapi.repository;
+
+import java.util.List;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import br.com.rocha.userapi.model.User;
+
+
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    User findByCpf(String cpf);
+
+    List<User> queryByNameLike(String name);
+}
