+++ user-api/src/main/resources/application.properties
@@ -1,0 +1,16 @@
+## Application port
+server.port = 8080
+
+##default connection pool
+spring.datasource.hikari.connectionTimeout=20000
+spring.datasource.hikari.maximumPoolSize=5
+
+##PostgreSQL
+spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
+spring.datasource.username=postgres
+spring.datasource.password=1234
+
+##Default Schema
+spring.flyway.schemas=users
+spring.jpa.properties.hibernate.default_schema=users
+
