+++ user-api/src/main/resources/db/migration/V1_create_user_table.sql
