+++ user-api/src/test/java/br/com/rocha/userapi/UserApiApplicationTests.java
@@ -1,0 +1,13 @@
+package br.com.rocha.userapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
